refactor: 딜러 입력 예외처리 PlayerNames 도메인으로 위임
