+++ src/main/java/domain/Name.java
@@ -8,19 +8,12 @@ public record Name(String name) {
     static final String DEALER_NAME_MESSAGE = "딜러라는 이름은 사용할 수 없습니다.";
 
     public Name {
-        validateNotDealerName(name);
         validateLength(name);
     }
 
     private void validateLength(String name) {
         if (name.length() < MIN_NAME_LENGTH || MAX_NAME_LENGTH < name.length()) {
             throw new IllegalArgumentException(NAME_LENGTH_MESSAGE);
-        }
-    }
-
-    private void validateNotDealerName(String name) {
-        if ("딜러".equals(name)) {
-            throw new IllegalArgumentException(DEALER_NAME_MESSAGE);
         }
     }
 }
