+++ src/test/java/domain/NameTest.java
@@ -1,14 +1,12 @@
 package domain;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import static domain.Name.DEALER_NAME_MESSAGE;
 import static domain.Name.NAME_LENGTH_MESSAGE;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 class NameTest {
 
@@ -27,13 +25,5 @@ class NameTest {
         assertThatThrownBy(() -> new Name(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(NAME_LENGTH_MESSAGE);
-    }
-
-    @DisplayName("플레이어 이름은 딜러가 될 수 없습니다.")
-    @Test
-    void validateNotDealerName() {
-        assertThatThrownBy(() -> new Name("딜러"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(DEALER_NAME_MESSAGE);
     }
 }
